♻️ refactor(core): expose node context types
