+++ packages/editor-shell/src/commands/clipboard/CutNodesCommand.ts
@@ -1,6 +1,6 @@
-import type { SceneNode } from "@design-editor/core"
+import type { NodeLocation, SceneNode } from "@design-editor/core"
 
-import type { Command, EditorReceiver, NodeLocation } from "../types"
+import type { Command, EditorReceiver } from "../types"
 
 interface RemovedClipboardNode {
 	location: NodeLocation
