+++ packages/editor-shell/src/commands/index.ts
@@ -2,7 +2,7 @@
 export { CommandHistory } from "./CommandHistory"
 export { CompositeCommand } from "./CompositeCommand"
 export { EditorReceiverImpl } from "./EditorReceiverImpl"
-export type { Command, EditorReceiver, MergableCommand, NodeLocation } from "./types"
+export type { Command, EditorReceiver, MergableCommand } from "./types"
 
 // ========== Node Commands ==========
 export * from "./node"
