+++ packages/editor-shell/src/commands/node/RemoveNodeCommand.ts
@@ -1,6 +1,6 @@
-import type { SceneNode } from "@design-editor/core"
+import type { NodeLocation, SceneNode } from "@design-editor/core"
 
-import type { Command, EditorReceiver, NodeLocation } from "../types"
+import type { Command, EditorReceiver } from "../types"
 
 /**
  * RemoveNodeCommand - 노드 삭제
