+++ packages/editor-shell/src/services/ClipboardRuntime.ts
@@ -1,7 +1,9 @@
+import type { NodeLocation } from "@design-editor/core"
+
 import { CutNodesCommand } from "../commands/clipboard/CutNodesCommand"
 import { PasteNodesCommand } from "../commands/clipboard/PasteNodesCommand"
 import type { CommandHistory } from "../commands/CommandHistory"
-import type { EditorReceiver, NodeLocation } from "../commands/types"
+import type { EditorReceiver } from "../commands/types"
 import { type ClipboardEntry, type ClipboardPayload, filterToTopLevelInPage, sortClipboardEntries } from "./clipboard"
 
 export class ClipboardRuntime {
